+++ feincms3/inline_ckeditor.py
@@ -1,5 +1,6 @@
 import json
 
+import django
 from django import forms
 from django.conf import settings
 from django.db import models
@@ -124,6 +125,8 @@ class InlineCKEditorWidget(forms.Textarea):
 
         attrs = kwargs.setdefault("attrs", {})
         attrs["data-inline-cke"] = id(self.config)
+        if django.VERSION < (4, 2):
+            attrs["data-inline-cke-dj41"] = True
         super().__init__(*args, **kwargs)
 
     @property